Added support for not overwriting certain files when updating.
<package version="1.1">
  <filegroup source="http://example.com/update/" dest=".">
    <file size="1024" version="any_string" overwrite="never">example.txt</file>
  </filegroup>
</package>

If the version of the file changes, but overwrite is disabled, the old version of the file will be stored in the update log. If the overwrite flag is later removed, the file will be updated.

+++ src/main/java/com/sk89q/mclauncher/model/Archive.java
@@ -54,6 +54,12 @@ public class Archive extends PackageFile {
                         File target = new File(getFile().getParentFile(), entry.getName());
                         checkSubchild(parent, target);
                         log.add(getFile(), target);
+
+                        // If we're not replacing, don't delete the existing file!
+                        if (getOverwrite() != null && target.exists()) {
+                            continue;
+                        }
+                        
                         target.getParentFile().mkdirs();
                         out = new BufferedOutputStream(
                                 new FileOutputStream(target));
@@ -97,7 +103,7 @@ public class Archive extends PackageFile {
     }
     
     /**
-     * Check to see whether the given file is within another give path.
+     * Check to see whether the given file is within another given path.
      * 
      * @param root folder that the child has to be under
      * @param child child file
